nf - Added resource management and AWS EC2 container management.
