--- repronim/orchestrator/__init__.py
@@ -1,16 +1,0 @@
-# emacs: -*- mode: python; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
-# ex: set sts=4 ts=4 sw=4 noet:
-# ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-#
-#   See COPYING file distributed along with the repronim package for the
-#   copyright and license terms.
-#
-# ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-"""Facility for installing and managing environments on targeted servers.
-
-"""
-
-__docformat__ = 'restructuredtext'
-
-# Allow base Orchestrator class to be available directly from the package.
-from repronim.orchestrator.base import Orchestrator
